Conflicts:
	katana2d/katana2d.iml

+++ app/src/main/java/com/toggle/flipped/PlayerInputSystem.java
@@ -16,7 +16,7 @@ public class PlayerInputSystem extends com.toggle.katana2d.System {
     }
 
     @Override
-    public void update(double dt) {
+    public void update(float dt) {
         TouchInputData touchData = mGame.getTouchInputData();
 
         for (Entity e: mEntities) {
@@ -39,9 +39,9 @@ public class PlayerInputSystem extends com.toggle.katana2d.System {
                     b.motionState = Bot.MotionState.MOVE;
                 }
 
-                // if vertical sliding direction is up, and sliding magnitude is more than 8
+                // if vertical sliding direction is up, and sliding magnitude is big enough
                 // then jump
-                if (touch.dy < -8)
+                if (touch.dy < -8 )
                     b.actionState = Bot.ActionState.JUMP_START;
             }
             else
